refactor: re-export session types and update imports
