+++ src/shared/entities/index.ts
@@ -2,3 +2,4 @@ export * from './base.types';
 export * from './base.enum';
 export * from './role.enum';
 export * from './rss.types';
+export * from './session.types';
